+++ transaction-history/translations.js
@@ -20,11 +20,16 @@ module.exports ={
         'ki': '99. Komeza',
         'sw': '99. Endelea'
     },
+    'prev_page': {
+        'en': '44. Back',
+        'ki': '44. Subira Inyuma',
+        'sw': '44. Rudi nyuma'
+    },
     'payment_detail': {
-        'en': 'Payment ID: $RepaymentId\nDate Received: $RepaymentDate\nSeason: $Season\nAmount: $Amount RwF\nPaid from: $PaidFrom',
-        'en-ke': 'Payment ID: $RepaymentId\nDate Received: $RepaymentDate\nSeason: $Season\nAmount: $Amount KES\nPaid from: $PaidFrom',
-        'ki': 'Nomero ndangagikorwa: $RepaymentId\nItariki yishyuriweho: $RepaymentDate\nIgihembwe: $Season\nAmafaranga: $Amount RwF\nYishyuwe avuye: $PaidFrom',
-        'sw': 'Rekodi ya malipo: $RepaymentId\nTarehe ya malipo: $RepaymentDate\nMsimu: $Season\nIdadi ya malipo: $Amount KES\nMalipo kutoka simu nambari: $PaidFrom',
+        'en': 'Payment ID: $RepaymentId\nDate Received: $RepaymentDate\nSeason: $Season\nAmount: $Amount RwF\nPaid from: $PaidFrom\n0) Back',
+        'en-ke': 'Payment ID: $RepaymentId\nDate Received: $RepaymentDate\nSeason: $Season\nAmount: $Amount KES\nPaid from: $PaidFrom\n0) Back',
+        'ki': 'Nomero ndangagikorwa: $RepaymentId\nItariki yishyuriweho: $RepaymentDate\nIgihembwe: $Season\nAmafaranga: $Amount RwF\nYishyuwe avuye: $PaidFrom\n0) Ahabanza',
+        'sw': 'Rekodi ya malipo: $RepaymentId\nTarehe ya malipo: $RepaymentDate\nMsimu: $Season\nIdadi ya malipo: $Amount KES\nMalipo kutoka simu nambari: $PaidFrom\n0) Rudi nyuma',
     },
     'invalid_last_4_nid_digits': {
         'en': 'You\'ve incorrectly entered the last four digits of your national ID. Please try again.',